refactoring mempool check
checking all the mempool transaction
will be needed for calculating the mempool
weight not only after boot, but also
at every major recheck and every new block.

Checking conflicts in incoming-transaction
of the mempool is needed only on boot.

We also refactor the order of the runSafe
condition checks - the this.checkMempool
case is relevant to handle only if
this.recheckMempoolTransactions is empty